+++ lib/talan/linear_counter.ex
@@ -1,4 +1,4 @@
-defmodule Talan.Counter do
+defmodule Talan.LinearCounter do
   @moduledoc """
   Linear probabilistic counter implementation with **safe concurrent access**,
   powered by the [:atomics](http://erlang.org/doc/man/atomics.html) module for cardinality estimation.
@@ -20,13 +20,13 @@ defmodule Talan.Counter do
   @type option :: {:hash_function, Talan.hash_function()}
   @type options :: list(option())
 
-  alias Talan.Counter
+  alias Talan.LinearCounter
   alias Talan.Validation
 
   @options [:hash_function]
 
   @doc """
-  Returns a new `%Talan.Counter{}` struct.
+  Returns a new `%Talan.LinearCounter{}` struct.
 
   `expected_cardinality` is the maximum number of unique items the counter will
   handle with an approximately 1% error rate.
@@ -39,11 +39,11 @@ defmodule Talan.Counter do
 
   ## Examples
 
-      iex> c = Talan.Counter.new(10_000)
-      iex> c |> Talan.Counter.put(["you", :can, Hash, {"any", "elixir", "term"}])
-      iex> c |> Talan.Counter.put("more")
-      iex> c |> Talan.Counter.put("another")
-      iex> c |> Talan.Counter.cardinality()
+      iex> c = Talan.LinearCounter.new(10_000)
+      iex> c |> Talan.LinearCounter.put(["you", :can, Hash, {"any", "elixir", "term"}])
+      iex> c |> Talan.LinearCounter.put("more")
+      iex> c |> Talan.LinearCounter.put("another")
+      iex> c |> Talan.LinearCounter.cardinality()
       3
   """
   @spec new(pos_integer()) :: t()
@@ -58,7 +58,7 @@ defmodule Talan.Counter do
     # Allocate ten bits per expected element, rounded up to a complete atomic word.
     required_size = div(expected_cardinality * 10 + 63, 64)
 
-    %Counter{
+    %LinearCounter{
       atomics_ref: :atomics.new(required_size, signed: false),
       filter_length: required_size * 64,
       hash_function: hash_function
@@ -75,13 +75,13 @@ defmodule Talan.Counter do
 
   ## Examples
 
-      iex> c = Talan.Counter.new(10_000)
-      iex> c |> Talan.Counter.put(["you", :can, Hash, {"any", "elixir", "term"}])
+      iex> c = Talan.LinearCounter.new(10_000)
+      iex> c |> Talan.LinearCounter.put(["you", :can, Hash, {"any", "elixir", "term"}])
       :ok
   """
   @spec put(t, any) :: :ok
   def put(
-        %Counter{
+        %LinearCounter{
           atomics_ref: atomics_ref,
           filter_length: filter_length,
           hash_function: hash_function
@@ -106,36 +106,36 @@ defmodule Talan.Counter do
 
   ## Examples
 
-      iex> counter = Talan.Counter.new(1000)
-      iex> Talan.Counter.put(counter, "Barna")
-      iex> Talan.Counter.clear(counter) == counter
+      iex> counter = Talan.LinearCounter.new(1000)
+      iex> Talan.LinearCounter.put(counter, "Barna")
+      iex> Talan.LinearCounter.clear(counter) == counter
       true
-      iex> Talan.Counter.cardinality(counter)
+      iex> Talan.LinearCounter.cardinality(counter)
       0
   """
   @spec clear(t()) :: t()
-  def clear(%Counter{atomics_ref: atomics_ref} = counter) do
+  def clear(%LinearCounter{atomics_ref: atomics_ref} = counter) do
     Abit.clear(atomics_ref)
     counter
   end
 
   @doc """
   Returns the estimated cardinality for the given
-  `%Talan.Counter{}` struct.
+  `%Talan.LinearCounter{}` struct.
 
   ## Examples
 
-      iex> c = Talan.Counter.new(10_000)
-      iex> c |> Talan.Counter.put(["you", :can, Hash, {"any", "elixir", "term"}])
-      iex> c |> Talan.Counter.put(["you", :can, Hash, {"any", "elixir", "term"}])
-      iex> c |> Talan.Counter.cardinality()
+      iex> c = Talan.LinearCounter.new(10_000)
+      iex> c |> Talan.LinearCounter.put(["you", :can, Hash, {"any", "elixir", "term"}])
+      iex> c |> Talan.LinearCounter.put(["you", :can, Hash, {"any", "elixir", "term"}])
+      iex> c |> Talan.LinearCounter.cardinality()
       1
-      iex> c |> Talan.Counter.put("more")
-      iex> c |> Talan.Counter.cardinality()
+      iex> c |> Talan.LinearCounter.put("more")
+      iex> c |> Talan.LinearCounter.cardinality()
       2
   """
   @spec cardinality(t) :: non_neg_integer
-  def cardinality(%Counter{atomics_ref: atomics_ref}) do
+  def cardinality(%LinearCounter{atomics_ref: atomics_ref}) do
     bit_count = Abit.bit_count(atomics_ref)
     set_bit_count = Abit.set_bits_count(atomics_ref)
     unset_bit_count = bit_count - set_bit_count